+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] — 2026-06-24 
+
+- The plugin interface is translated into 11 languages
+- Added link to manage saved linguistic Instructions 
+
+## [1.1.2] — 2026-06-16
+
+- Added support json, yaml, other plain-text formats
+
 ## [1.1.1] — 2026-06-15
 
 - Set correct min API version